requirements.txt
allure-python-commons==2.12.0
appdirs==1.4.4
assertpy==1.1
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
distlib==0.3.6
filelock==3.9.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pipenv==2022.12.19
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.2.1
pytest-ordering==0.6
requests==2.28.2
selenium==4.8.0
six==1.16.0
toml==0.10.2
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
webdriver-manager==3.8.5
xlrd==2.0.1
gitPython==3.1.30
pytest-rerunfailures==11.0
pytest-testconfig>=0.2.0
pytest-logger>=0.5.1
marshmallow>=3.7.1
Faker==16.6.0
uuid~=1.30
random-username==1.0.2
python-dateutil~=2.8.2
boto3==1.26.55
pytest-bdd==6.1.1
pytest-bdd-ng==1.2.3
allure-pytest==2.12.0
python-dotenv==0.21.1